Move graph_shortest_path to utils/graph.py
