+++ sklearn/utils/graph.py
@@ -12,6 +12,8 @@ sparse matrices.
 import numpy as np
 from scipy import sparse
 
+from graph_shortest_path import graph_shortest_path
+
 
 ###############################################################################
 # Path and connected component analysis.
